refactor: implemented the gifts store provider on portal layout tsx and implemented useGiftsStore on useGetGifts hook
