+++ frontend/src/app/portal/giftList/[id]/(components)/conditionalRender/conditionalRender.tsx
@@ -4,15 +4,15 @@ import GuestList from '@/app/portal/giftList/[id]/(components)/giftsListDisplay/
 import GuestRequest from '@/app/portal/giftList/[id]/(components)/giftsListDisplay/guest-request'
 import OwnerList from '@/app/portal/giftList/[id]/(components)/giftsListDisplay/owner-list'
 import { useContextWrap } from '@/contextAPI/context'
-import useGetData from '@/app/portal/giftList/[id]/(hooks)/useGetData'
 import { useEffect } from 'react'
+import useGetGifts from '@/app/portal/giftList/[id]/(hooks)/useGetGifts'
 
 export default function ConditionalRenderingListPage() {
   const { notGuest, isCreator } = useContextWrap()
-  const getData = useGetData()
+  const getGifts = useGetGifts()
 
   useEffect(() => {
-    getData()
+    getGifts()
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
   return (
